requirements.txt
opencv-python
speechrecognition
pyttsx3
requests
wikipedia
deepface
python-dotenv
google-generativeai

 